Fixed bugs added a new room and new upgrade
Added a spike trap room which has no animation right now but deals damage to the player and IS noticeable in the ceiling. Also added a stamina regen upgrade which upgrades the speed the stamina regenerates.

I have also made it so changing the FPS is more noticeable and so that buying upgrades of exact prices should work now.

Finally I have fixed a bug where if you died the timer would keep running.